+++ custom_components/tapo_event_bridge/const.py
@@ -2,7 +2,7 @@
 
 DOMAIN = "tapo_event_bridge"
 NAME = "Tapo Event Bridge"
-VERSION = "1.3.0"
+VERSION = "1.4.0"
 
 PLATFORMS = ("binary_sensor", "button", "number", "select", "sensor", "switch")
 EVENT_CAMERA = f"{DOMAIN}_event"